+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Releases
 
+## pandoc-utils 0.5.1 (2020-5-24)
+
+- Fix the link to Hakyll in package description.
+
+## pandoc-utils 0.5.0 (2020-5-24)
+
+- Reorganize documentations and API.
+- Add new functions `convertFilter` and `convertFilterM` to convert between
+  filter functions.
+- Enable implicit conversion in `getConcatedFilter` and `getConcatedFilterM`.
+- Add a module exporting all the utility functions provided by the package.
+
 ## pandoc-utils 0.4.0 (2020-5-24)
 
 - Initial public release.